Fix test in full dotnet
